add s21_ceil() with tests

+++ include/s21_math.h
@@ -19,7 +19,7 @@ int s21_abs(int x);
 long double s21_acos(double x);
 long double s21_asin(double x);
 long double s21_atan(double x);
-long double s21_ceil(double x);
+double s21_ceil(double x);
 long double s21_cos(double x);
 long double s21_exp(double x);
 double s21_fabs(double x);
